now compiles on della, small changes to workflow
+ static_assert(false) statements removed
+ now properly calling template ContainerType in interface_container.hpp
- removed call to reference-build's xmeshfem in cg_compare
+ added a "if dir exists" check in build session before shutil.rmtree

+++ workflow_int_ptr_ptr/noxfile.py
@@ -16,7 +16,8 @@ nox.options.reuse_existing_virtualenvs = True
 def reconfigure_build(session):
     session.cd(config.get("specfem.root"))
     build_dir = config.get("specfem.live.build")
-    shutil.rmtree(build_dir)
+    if os.path.exists(build_dir):
+        shutil.rmtree(build_dir)
     os.system(
         f"cmake -S . -B {build_dir} {config.get('specfem.live.cmake_build_options')}"
     )
